pyproject.toml
[project]
name = "scc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.65",
    "bertopic>=0.17.4",
    "gliner>=0.2.25",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "openai>=2.21.0",
    "pandas>=3.0.1",
    "pip>=26.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "spacy>=3.8.11",
]
